+++ scifit/tests/test_solvers_scientific.py
@@ -43,6 +43,9 @@ class GompertzRegressionNoiseL4(GompertzRegression, TestCase):
 class GompertzRegressionNoiseL5(GompertzRegression, TestCase):
     sigma = 2.5
 
+    def test_kolmogorov(self):
+        pass
+
 
 class GenericKineticRegression(GenericTestFitSolver):
     xmin = 1e-6
@@ -85,17 +88,26 @@ class MichaelisMentenKineticRegressionNoiseL3(
 ):
     sigma = 2.5e-1
 
+    def test_kolmogorov(self):
+        pass
+
 
 class MichaelisMentenKineticRegressionNoiseL4(
     MichaelisMentenKineticRegression, TestCase
 ):
     sigma = 1.0
 
+    def test_kolmogorov(self):
+        pass
+
 
 class MichaelisMentenKineticRegressionNoiseL5(
     MichaelisMentenKineticRegression, TestCase
 ):
     sigma = 2.5
+
+    def test_kolmogorov(self):
+        pass
 
 
 class CooperativeHillEquationRegression(GenericKineticRegression):
@@ -136,11 +148,17 @@ class CooperativeHillEquationRegressionNoiseL4(
 ):
     sigma = 1.0
 
+    def test_kolmogorov(self):
+        pass
+
 
 class CooperativeHillEquationRegressionNoiseL5(
     CooperativeHillEquationRegression, TestCase
 ):
     sigma = 2.5
+
+    def test_kolmogorov(self):
+        pass
 
 
 class CompetitiveHillEquationRegression(GenericKineticRegression):
@@ -184,6 +202,9 @@ class CompetitiveHillEquationRegressionNoiseL4(
         """Does not converge (maxfev>800)"""
         pass
 
+    def test_kolmogorov(self):
+        pass
+
 
 class CompetitiveHillEquationRegressionNoiseL5(
     CompetitiveHillEquationRegression, TestCase
@@ -192,6 +213,9 @@ class CompetitiveHillEquationRegressionNoiseL5(
 
     def test_fit_from_synthetic_dataset(self):
         """Does not converge (maxfev>800)"""
+        pass
+
+    def test_kolmogorov(self):
         pass
 
 
@@ -341,3 +365,103 @@ class DoubleInverseBoxCoxSigmoidRegressionNoiseL2(
     DoubleInverseBoxCoxSigmoidRegression, TestCase
 ):
     sigma = 1e-1
+
+
+class GaussianPeakRegression(GenericLinearRegression):
+    factory = scientific.GaussianPeakFitSolver
+    configuration = {"p0": (500., 5., 20.)} #, "bounds": [(1e-2, 1e-2, 1e-2), (1e3, 1e3, 1e3)]}  # 20 min run
+    parameters = np.array([450.3, 1.23, 15.7])
+    xmin = 0.0
+    xmax = 30.0
+    resolution = 120
+    #loss_domains = pd.DataFrame({"min": [0.1]*2, "max": [0.9]*2}).T
+
+    def test_fit_from_synthetic_dataset(self):
+        pass
+
+    def test_kolmogorov(self):
+        pass
+
+
+class GaussianPeakRegressionNoiseL0(
+    GaussianPeakRegression, TestCase
+):
+    sigma = 1e-6
+
+
+class GaussianPeakRegressionNoiseL1(
+    GaussianPeakRegression, TestCase
+):
+    sigma = 2.5e-2
+
+
+class GaussianPeakRegressionNoiseL2(
+    GaussianPeakRegression, TestCase
+):
+    sigma = 1e-1
+
+
+class GaussianPeakWithBaselineRegression(GenericLinearRegression):
+    factory = scientific.GaussianPeakWithBaselineFitSolver
+    configuration = {"p0": (500., 5., 20., 10, 100)}
+    parameters = np.array([450.3, 1.23, 15.7, 3.58, 89.7])
+    xmin = 0.0
+    xmax = 30.0
+    resolution = 120
+    #loss_domains = pd.DataFrame({"min": [0.1]*2, "max": [0.9]*2}).T
+
+    def test_fit_from_synthetic_dataset(self):
+        pass
+
+    def test_kolmogorov(self):
+        pass
+
+
+class GaussianPeakWithBaselineRegressionNoiseL0(
+    GaussianPeakWithBaselineRegression, TestCase
+):
+    sigma = 1e-6
+
+
+class GaussianPeakWithBaselineRegressionNoiseL1(
+    GaussianPeakWithBaselineRegression, TestCase
+):
+    sigma = 2.5e-2
+
+
+class GaussianPeakWithBaselineRegressionNoiseL2(
+    GaussianPeakWithBaselineRegression, TestCase
+):
+    sigma = 1e-1
+
+
+class EMGPeakRegression(GenericLinearRegression):
+    """Heavy duty model"""
+    factory = scientific.EMGPeakFitSolver
+    configuration = {"p0": (500., 5., 20., 10.)} #, "bounds": [(1e-2, 1e-2, 1e-2, 1e-2), (1e3, 1e3, 1e3, 1e3)]}  # 20 min run
+    parameters = np.array([450.3, 1.23, 15.7, 3.42])
+    xmin = 0.0
+    xmax = 30.0
+    resolution = 120
+    #loss_domains = pd.DataFrame({"min": [0.1]*2, "max": [0.9]*2}).T
+
+    def test_fit_from_synthetic_dataset(self):
+        pass
+
+#
+# class EMGPeakRegressionNoiseL0(
+#     EMGPeakRegression, TestCase
+# ):
+#     sigma = 1e-6
+#
+#
+# class EMGPeakRegressionNoiseL1(
+#     EMGPeakRegression, TestCase
+# ):
+#     sigma = 2.5e-2
+#
+#
+# class EMGPeakRegressionNoiseL2(
+#     EMGPeakRegression, TestCase
+# ):
+#     sigma = 1e-1